eatgo-customer-api/src/main/java/com/reservationsystem/application/UserService.java
package com.reservationsystem.application;

import com.reservationsystem.domain.User;
import com.reservationsystem.domain.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.Optional;


@Service
public class UserService {


    private UserRepository userRepository;

    private PasswordEncoder passwordEncoder;

    @Autowired
    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder){
        this.userRepository= userRepository;
        this.passwordEncoder=passwordEncoder;
    }


    public User registerUser(String email, String name, String password) {
        Optional<User>existed = userRepository.findByEmail(email);
        if(existed.isPresent()){
            throw new EmailExistedException(email);
        }

        String encodedPassword = passwordEncoder.encode(password);

        User mockUser = User.builder()
                .email(email)
                .name(name)
                .password(encodedPassword)
                .level(1L)
                .build();

        return userRepository.save(mockUser);
    }


    public User authenticate(String email, String password) {
        User user = userRepository.findByEmail(email)
                .orElseThrow(
                        ()-> new EmailExistedException(email));

        if(!passwordEncoder.matches(password, user.getPassword())){
            throw new PasswordWrongException();
        }

        return user;
    }
}
